extract selector for LoaderScheduleRow
