+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/schedule/LoaderSchedulesDelegatee.scala
@@ -9,7 +9,7 @@ import x7c1.linen.glue.res.layout.SettingScheduleLayout
 import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.glue.service.ServiceLabel.Updater
 import x7c1.linen.repository.channel.subscribe.SubscribedChannel
-import x7c1.linen.repository.loader.schedule.{ChannelLoaderSchedule, PresetLoaderSchedule, TimeRange}
+import x7c1.linen.repository.loader.schedule.LoaderScheduleRow
 import x7c1.linen.scene.updater.UpdaterMethods
 import x7c1.wheat.lore.resource.AdapterDelegatee
 import x7c1.wheat.macros.intent.{IntentExpander, ServiceCaller}
@@ -42,34 +42,26 @@ class LoaderSchedulesDelegatee (
     helper.close()
   }
   def setupFor(accountId: Long): Unit = {
-    layout.schedules setLayoutManager new LinearLayoutManager(activity)
-    layout.schedules setAdapter new ScheduleRowAdapter(
-      delegatee = AdapterDelegatee.create(scheduleRowProviders, dummySchedules),
-      providers = timeRowProviders,
-      onMenuSelected = showMenu(accountId)
-    )
-  }
-  lazy val dummySchedules = Sequence from Seq(
-    PresetLoaderSchedule(
-      scheduleId = 111,
-      name = "Load subscribed channels at..",
-      enabled = true,
-      startRanges = Sequence from Seq(
-        TimeRange(3, 4),
-        TimeRange(9, 10),
-        TimeRange(15, 16)
-      )
-    )
-  ) ++ createChannelSchedules
-
-  private def createChannelSchedules = {
-    (0 to 20) map { n =>
-      ChannelLoaderSchedule(
-        scheduleId = 111 * n,
-        name = s"Load channel : $n",
-        enabled = true
-      )
+    helper.selectorOf[LoaderScheduleRow] collectBy accountId match {
+      case Right(schedules) =>
+        val adapter = createAdapter(accountId, schedules)
+        layout.schedules setLayoutManager new LinearLayoutManager(activity)
+        layout.schedules setAdapter adapter
+      case Left(e) =>
+        Log error format(e){"[failed]"}
     }
+  }
+  private def createAdapter[A: AccountIdentifiable]
+    (account: A, schedules: Sequence[LoaderScheduleRow]) = {
+
+    new ScheduleRowAdapter(
+      delegatee = AdapterDelegatee.create(
+        providers = scheduleRowProviders,
+        sequence = schedules
+      ),
+      providers = timeRowProviders,
+      onMenuSelected = showMenu(account)
+    )
   }
   private def showMenu[A: AccountIdentifiable]
     (account: A)(event: ScheduleSelected) = {
